Fix a bizarre bug, where the rendertime is always printed as "0.0"
It seems that when you change the printing code to print float (instead
of double), the problem disappears, so this is what the fix does.

+++ src/sdl.cpp
@@ -69,7 +69,7 @@ void displayVFB(Color vfb[VFB_MAX_SIZE][VFB_MAX_SIZE])
 	SDL_Flip(screen);
 }
 
-void setWindowCaption(const char* msg, double renderTime)
+void setWindowCaption(const char* msg, float renderTime)
 {
 	if (renderTime >= 0) {
 		char message[128];
